requirements.txt
pandas
numpy
scikit-learn
joblib
streamlit
matplotlib
seaborn
psycopg2-binary
